+++ kernel/src/stddef.h
@@ -1,5 +1,22 @@
-#ifndef _WOS_STDDEF_H
-#define _WOS_STDDEF_H
+/*
+Copyright (©) 2022-2023  Frosty515
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
+#ifndef _STDDEF_H
+#define _STDDEF_H
 
 typedef unsigned long size_t;
 
@@ -21,4 +38,4 @@ typedef struct {
 
 #define offsetof(t, d) __builtin_offsetof(t, d)
 
-#endif /* _WOS_STDDEF_H */
+#endif /* _STDDEF_H */