+++ kernel/src/stdio.hpp
@@ -1,5 +1,22 @@
-#ifndef _KERNEL_STDIO_HPP
-#define _KERNEL_STDIO_HPP
+/*
+Copyright (©) 2022-2023  Frosty515
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
+#ifndef _STDIO_HPP
+#define _STDIO_HPP
 
 #include <stddef.h>
 #include <stdint.h>
@@ -21,4 +38,4 @@ void vfprintf(const fd_t file, const char* format, va_list args);
 
 void fwrite(const void* ptr, const size_t size, const size_t count, const fd_t file);
 
-#endif /* _KERNEL_STDIO_HPP */
+#endif /* _STDIO_HPP */